@@ -1,12 +1,12 @@
 [tool.poetry]
 name = "eapix"
-version = "0.8.0"
+version = "0.9.0"
 description = ""
 authors = ["Jesse Mather <jmather@arista.com>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.8.1"
+python = "^3.10"
 httpx = "^0.27"
 click = "^8.1"
 
