@@ -5,36 +5,32 @@
 from typing import List, Optional, Tuple, Union
 from dataclasses import dataclass
 
-from eapix.environments import EAPI_DEFAULT_FORMAT
+from eapix.environment import EAPI_DEFAULT_FORMAT
+
+Auth = Tuple[str, str]
+
+Certificate = Union[str, Tuple[str, str], Tuple[str, str, str]]
 
 @dataclass
-class BaseType:
-    pass
-
-@dataclass
-class Error(BaseType):
-    code: int
-    message: str
-
-@dataclass
-class Command(BaseType):
+class Command:
     cmd: str
-    input: str = ""
+    input: Optional[str] = None
+
+CommandList = List[Union[str, tuple[str, str], Command]]
 
 @dataclass
-class EapiOptions(BaseType):
+class EapiOptions:
     version: int = 1
-    format: str = EAPI_DEFAULT_FORMAT
+    encoding: str = EAPI_DEFAULT_FORMAT
     timestamps: Optional[bool] = None
     auto_complete: Optional[bool] = None
     expand_aliases: Optional[bool] = None
     include_error_detail: Optional[bool] = None
     streaming: Optional[bool] = None
 
-CommandList = List[Union[str, Command]]
-
-Auth = Tuple[str, Optional[str]]
-
-Certificate = Optional[Union[str, Tuple[str, str], Tuple[str, str, str]]]
+@dataclass
+class Error:
+    code: int
+    message: str
 
 Timeout = Union[None, float, Tuple[float, float, float, float]]
